Revert "refactor: remove InspectableWebContentsViewMac in favor of the Views version (#41326)"
This reverts commit e67ab9a93dadccecff30de50ab4555191c30b6c4.
