+++ shell/browser/ui/cocoa/electron_ns_window.mm
@@ -200,11 +200,6 @@ void SwizzleSwipeWithEvent(NSView* view, SEL swiz_selector) {
 }
 
 - (NSRect)constrainFrameRect:(NSRect)frameRect toScreen:(NSScreen*)screen {
-  // We initialize the window in headless mode to allow painting before it is
-  // shown, but we don't want the headless behavior of allowing the window to be
-  // placed unconstrained.
-  self.isHeadless = false;
-
   // Resizing is disabled.
   if (electron::ScopedDisableResize::IsResizeDisabled())
     return [self frame];
